Break out some stuff into a different file. main is getting kinda large.
